Create Kiruthika_infix_to_postfix.cpp
Kiruthika is working on a programming project that requires the conversion of infix mathematical expressions into postfix notation. To accomplish this task efficiently, she seeks your assistance in developing a program.



Your goal is to create a program that accepts an infix expression as input and converts it into postfix notation. Kiruthika is interested in a program that can handle various arithmetic operations, parentheses, and operands within the input expression.



Note: This is a sample question asked in a Deloitte interview.

Input format :
The input consists of the infix expression and arithmetic operators (+, -, *, /), and parentheses () with a single space.

Output format :
The output prints the postfix expression of the input expression without any space.

Code constraints :
The input includes (, ), +, -, * / and upper alphabets.

Sample test cases :
Input 1 :
(A + (B * C))
Output 1 :
ABC*+
Input 2 :
 A / D - E
Output 2 :
AD/E-