Create double-ended_queue.cpp
Your task is to implement a double-ended queue data structure. A deque is a linear data structure that allows elements to be inserted or removed from both the front and the rear. 



Your program should provide the following functionality:

Initialize an empty deque.
Push an element to the back of the deque.
Check if the deque is empty.
Find and print the maximum element in the deque.
Input format :
The input consists of the elements that should be inserted into the deque.

The input is terminated by entering -1.

Output format :
The output prints the maximum element in the given deque.

Sample test cases :
Input 1 :
1
2
3
4
-1
Output 1 :
4
Input 2 :
41
36
-1
Output 2 :
41