Create mergeSort.cpp
You have a dictionary with a collection of characters. Implement a program that sorts the characters in lexicographic order using the recursive Merge Sort algorithm.



Call the function mergeSort to sort the array of characters. The program should take the unsorted list of characters as input and output the sorted list of characters.



Write a logic for merge sorting and a recursive function.



Note: The lexicographical order of characters will be 'A', 'B', 'C', …, 'Y', 'Z', 'a', 'b', 'c', …, 'y', 'z'.

Input format :
The first line of input consists of the characters n, representing the number of characters in the dictionary.

The second line of input consists of n characters, separated by spaces.

Output format :
The output displays a sorted list of characters in lexicographic order, with each character separated by a space.



Refer to the sample output for the formatting specifications.

Code constraints :
MAX_Char_Length 30

1 ≤ n ≤ 30

Input characters should be lowercase or uppercase.

Sample test cases :
Input 1 :
4
z g v a
Output 1 :
Sorted Characters: a g v z 
Input 2 :
3
a b a
Output 2 :
Sorted Characters: a a b 
Input 3 :
5
a b A c D
Output 3 :
Sorted Characters: A D a b c