Create logistics_company_Quick-Sort.cpp
Imagine you are a software developer working for a logistics company. Your task is to create a program that can efficiently sort a list of packages based on their weights using the Quick-Sort algorithm. Each package has a weight value associated with it.



The program should take an array of package objects as input, where each object contains the weight of a package. The program should then sort the array in ascending order based on the weights of the packages.

Input format :
The first line of input consists of an integer N, representing the number of packages.

The second line consists of N integers, representing the weight of the packages, separated by space.

Output format :
The output displays N integers representing the weights of the packages after they have been sorted in ascending order.

Code constraints :
N > 0

weight > 0

Sample test cases :
Input 1 :
4
15 25 35 10
Output 1 :
10 15 25 35 
Input 2 :
3
120 60 80
Output 2 :
60 80 120