Create AddAmountBase.cpp
You are developing a simple banking application that allows users to manage their account balance. To implement this, you have two classes: "AddAmountBase" and "AddAmount". The "AddAmountBase" class serves as the base class, which contains a protected data member amount to store the current balance. It has two constructors: one initializes the balance to 50, and another adds a given amount, 'a' to the balance.


Your task is to create a derived class "AddAmount" that inherits from the "AddAmountBase" class. The "AddAmount" class should have two constructors, one initializes the balance to 50, and another adds a given amount to the balance using the base class constructors.


The "AddAmount" class should also have a member function named print_amount() to display the current balance.
Input format :

The input consists of the amount in the piggy bank.
Output format :

The output prints the total amount in Piggie Bank after adding the input to the initial amount.
Code constraints :

0<=amount<=100000
Sample test cases :
Input 1 :

400

Output 1 :

450

Input 2 :

250

Output 2 :

300